requirements.txt
nltk==3.5
tqdm==4.48.2
contractions==0.0.25
gensim==3.8.3
scikit-learn==0.23.2
matplotlib==3.2.2
plotly==4.10.0
python-graphviz
pydot